requirements.txt
aiofiles==0.4.0
aiohttp==3.5.4
async-timeout==3.0.1
attrs==19.1.0
beautifulsoup4==4.7.1
blessed==1.15.0
bs4==0.0.1
certifi==2019.3.9
chardet==3.0.4
enlighten==1.3.0
fire==0.1.3
h11==0.8.1
h2==3.1.0
hpack==3.0.0
http3==0.6.7
httpcore==0.2.1
hyperframe==5.2.0
idna==2.8
importlib-metadata==0.9
multidict==4.5.2
path.py==12.0.1
python-slugify==3.0.2
requests==2.21.0
requests-async==0.4.1
rfc3986==1.3.2
six==1.12.0
soupsieve==1.9.1
text-unidecode==1.2
unicode-slugify==0.1.3
Unidecode==1.1.1
urllib3==1.24.3
wcwidth==0.1.7
yarl==1.3.0
zipp==0.4.0
